+++ tests/mocks/app/routes/app1/router.js
@@ -1,4 +1,4 @@
-const service = require('../index.js');
+const service = require('../../index.js');
 
 const router = service.appManager.get('app1').buildRouter({
     version: 1.0,