+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## FUTURE
+## v1.5.0
 
 * [ADDED] - `ValidationError.API_CODE` static writable property (which defaults to `validationFailure`) - sets `api_code` value for validation error http 400 response
 * [ADDED] - Ajv option `allErrors=true` is now supported. if enabled, validation response will include a collection of failures
@@ -6,7 +6,7 @@
 * [FIXED] - shell commands should print a short description of failure reason to stderr
 * [ADDED] - http `Route` implementation supports a regular expression as `url` option value
 
-## 1.4.0
+## v1.4.0
 
 * [ADDED] - `common/Route.prototype.acceptedContentTypes()` which returns a collection of supported request content mime types
 * [ADDED] - configurable `cluster` option of `bi-service run --cluster=<number>` where `<number>` can be a percentage amount (from number of available cpu threads) of childs to be forked in the case of floating point value, or exlicit number of childs when integer is provided